Revert "Possible move of functionality into DemographicParity"
This reverts commit 9baa9502f8c92176b0d120a1383c3d49aea3f562.
